requirements.txt
# Core dependencies
rich==14.3.2
llama-index-core>=0.10.0
llama-index-llms-ollama==0.9.1

# Optional but recommended
tiktoken>=0.5.0
pypdf>=3.17.0
chromadb>=0.4.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
